confirms all abbrev. letters used in full term
updates to confirm that all letters in the abbreviation are found in
the full term. updates existing tests to account for this. streamlines
.travis.yml file.

fixes typo in .travis.yml

confirms all abbrev. letters used in full term

updates to confirm that all letters in the abbreviation are found in
the full term. updates existing tests to account for this. streamlines
.travis.yml file