Use different start_idx for matches and non matches.
When utils.replace loops through the words in the text, it either finds
the abbreviation or it doesn’t.  If it doesn’t find the abbreviation,
the next step should look after the end of the checked word. If it does
find the abbreviation, the next step should look after the length of
the full term that was sandwiched in there. That’s the key here, I
think.